+++ ignite/cmd/generate_vuex.go
@@ -40,10 +40,7 @@ func generateVuexHandler(cmd *cobra.Command, _ []string) error {
 		return err
 	}
 
-	output, err := cmd.Flags().GetString(flagOutput)
-	if err != nil {
-		return err
-	}
+	output, _ := cmd.Flags().GetString(flagOutput)
 
 	var opts []chain.GenerateTarget
 	if flagGetEnableProtoVendor(cmd) {